Make upbound-agent owned by versions ConfigMap
Updates the upbound-bootstrapper to make upbound-agent owned by the
versions ConfigMap, which serves as the official indicator of installed
versions in a UXP install. This means that uninstalling UXP will now
successfully clean up the agent, as will deleting the control plane
token secret, which will cause upbound-bootstrapper to manually delete
the upbound-agent deployment.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>