Fix multi-platform builds for arm64
Currently we're somehow building a multi-platform image (i.e. manifest
list) but the arm64 image has an amd64 bootstrapper binary in it. I'm
honestly not exactly sure how this is happening.

In an attempt to fix it, I've copied the Makefile and Dockerfile from
Crossplane v1.16 (before it moved to Earthly for image builds). I've
verified that the arm64 image works locally - I think we need to
actually build and push an image in CI to test multiplatform.

Signed-off-by: Nic Cope <nicc@rk0n.org>